more help stuff, fixing proper start state

+++ trunk/seeautz/source/glutFramework/glutFramework/helpScreenState.h
@@ -30,7 +30,7 @@ private:
 	int textOffsetX;
 	int textOffsetY;
 	int offsetAmt;
-	int textSpacing;
+	double textSpacing;
 	int textSpacing2;
 
 	// tile stuff
